SR-121: ODK config moved out as a separate model
