--- g2p_odk_importer/data/odk_cron.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-
-    <!-- This cron job will run every day at midnight -->
-    <!-- <record id="ir_cron_odk_import" model="ir.cron">
-        <field name="name">ODK Import</field>
-        <field name="user_id" ref="base.user_root"/>
-        <field name="interval_number">1</field>
-        <field name="interval_type">minutes</field>
-        <field name="numbercall">-1</field>
-        <field name="doall" eval="False"/>
-        <field name="active" eval="True"/>
-        <field name="model_id" ref="model_odk_config"/>
-        <field name="code">model_odk_config.import_records()</field>
-    </record> -->
-
-</odoo>
