+++ g2p_odk_importer/models/__init__.py
@@ -2,3 +2,4 @@
 
 from . import odk_client
 from . import odk_config
+from . import odk_import
