+++ g2p_odk_importer/tests/__init__.py
@@ -1,4 +1,4 @@
 # Part of Newlogic G2P. See LICENSE file for full copyright and licensing details.
 
 from . import test_odk_client
-from . import test_odk_config
+from . import test_odk_import
