+++ g2p_odk_importer/views/odk_config_views.xml
@@ -1,13 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <record id="action_odk_conf_action" model="ir.actions.server">
-        <field name="name">ODK Config Start</field>
-        <field name="model_id" ref="model_odk_config" />
-        <field name="state">code</field>
-        <field name="code">
-            record.odk_import_action_trigger()
-        </field>
-    </record>
 
     <record id="view_odk_config_tree" model="ir.ui.view">
         <field name="name">view_odk_config_tree</field>
@@ -17,28 +9,6 @@
             <tree>
                 <field name="name" />
                 <field name="base_url" />
-                <field name="job_status" />
-                <button
-                    type="action"
-                    name="%(action_odk_conf_action)d"
-                    invisible="job_status != 'draft'"
-                    string="Start"
-                    title="Action odk conf action"
-                />
-                <button
-                    type="action"
-                    name="%(action_odk_conf_action)d"
-                    invisible="job_status not in ('started','running')"
-                    string="Stop"
-                    title="Action odk conf action"
-                />
-                <button
-                    type="action"
-                    name="%(action_odk_conf_action)d"
-                    invisible="job_status != 'completed'"
-                    string="Restart"
-                    title="Action odk conf action"
-                />
             </tree>
         </field>
     </record>
@@ -47,72 +17,13 @@
         <field name="name">odk.config.form</field>
         <field name="model">odk.config</field>
         <field name="arch" type="xml">
-            <form string="ODK Configuration">
-                <header>
-                    <button
-                        name="test_connection"
-                        string="Test Connection"
-                        type="object"
-                        class="oe_highlight"
-                        titile="Test connection"
-                    />
-                    <button
-                        name="import_records"
-                        string="Import Records"
-                        type="object"
-                        class="oe_highlight"
-                        title="Import records"
-                    />
-                </header>
-
+            <form>
                 <sheet>
-                    <div class="oe_button_box" name="button_box">
-                            <button
-                            type="action"
-                            name="%(action_odk_conf_action)d"
-                            icon="fa-play"
-                            invisible="job_status != 'draft'"
-                            title="Action odk conf action"
-                        >
-                                <span class="o_stat_text">Start</span>
-                            </button>
-                            <button
-                            type="action"
-                            name="%(action_odk_conf_action)d"
-                            icon="fa-stop"
-                            invisible="job_status not in ('started','running')"
-                            title="Action odk conf action"
-                        >
-                                <span class="o_stat_text">Stop</span>
-                            </button>
-                            <button
-                            type="action"
-                            name="%(action_odk_conf_action)d"
-                            icon="fa-repeat"
-                            invisible="job_status != 'completed'"
-                            title="Action odk conf action"
-                        >
-                                <span class="o_stat_text">Restart</span>
-                            </button>
-                        </div>
-                    <group string="ODK Central">
+                    <group string="ODK Configuration">
                         <field name="name" />
                         <field name="base_url" />
                         <field name="username" />
-                        <field name="password" password="True" />
-                        <field name="job_status" invisible="True" />
-                        <field name="last_sync_time" />
-                    </group>
-                    <group string="Project details">
-                        <field name="project" />
-                        <field name="form_id" />
-                    </group>
-                    <group string="Target settings">
-                        <field name="target_registry" />
-                        <field name="json_formatter" />
-                    </group>
-                    <group string="Time interval">
-                        <field name="interval_hours" />
+                        <field name="password" password="true" />
                     </group>
                 </sheet>
             </form>
@@ -124,15 +35,7 @@
         <field name="type">ir.actions.act_window</field>
         <field name="res_model">odk.config</field>
         <field name="view_mode">tree,form</field>
-        <!-- <field name="search_view_id" ref="odk_connector.view_odk_config_search"/> -->
         <field name="context">{}</field>
     </record>
-
-    <!-- <record id="menu_odk_config" model="ir.ui.menu">
-        <field name="name">ODK Configuration</field>
-        <field name="parent_id" ref="base.menu_configuration"/>
-        <field name="action" ref="odk_connector.action_odk_config"/>
-    </record> -->
-
 
 </odoo>
