+++ g2p_odk_importer/views/odk_menu.xml
@@ -3,12 +3,19 @@
 
     <menuitem id="odk_menu_root" name="ODK" sequence="30" active="True" />
     <menuitem
-        id="odk_menu_config"
-        name="Configuration"
-        active="True"
+        id="odk_import_menu"
+        name="ODK Import"
+        parent="odk_menu_root"
+        action="g2p_odk_importer.action_odk_import"
+        sequence="1"
+    />
+
+    <menuitem
+        id="odk_config_menu"
+        name="ODK Configuration"
         parent="odk_menu_root"
         sequence="1"
-        action="action_odk_config"
+        action="g2p_odk_importer.action_odk_config"
     />
 
 </odoo>
